+++ lib/unix_representations.ml
@@ -11,7 +11,9 @@ external nativeint_of_dir_handle : Unix.dir_handle -> nativeint
 external dir_handle_of_nativeint : nativeint -> Unix.dir_handle
   = "dir_handle_of_nativeint_stub"
 
+external dir_handle_clean : Unix.dir_handle -> unit
+  = "dir_handle_clean_stub"
+
 let int_of_file_descr : Unix.file_descr -> int = Obj.magic
 
 let file_descr_of_int : int -> Unix.file_descr = Obj.magic
-