+++ lib/unix_type_representation_stubs.c
@@ -10,7 +10,6 @@
 #include <caml/memory.h>
 #include <caml/alloc.h>
 #include <caml/unixsupport.h>
-#include <caml/unixsupport.h>
 
 #ifdef HAS_DIRENT
 #include <dirent.h>
@@ -26,7 +25,14 @@ value nativeint_of_dir_handle_stub(value dir_handle)
 value dir_handle_of_nativeint_stub(value nativeint)
 {
   DIR *d = (DIR *)(intptr_t)Nativeint_val(nativeint);
-  value res = alloc_small(1, Abstract_tag);
+  value res = caml_alloc_small(1, Abstract_tag);
   DIR_Val(res) = d;
   return res;
+}
+
+void dir_handle_clean_stub(value dir_handle)
+{
+  CAMLparam1(dir_handle);
+  DIR_Val(dir_handle) = (DIR *) NULL;
+  CAMLreturn0;
 }